modules/transport/requirements.txt
# FastAPI et serveur ASGI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Base de données PostgreSQL
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlalchemy==2.0.36

# MQTT et WebSocket pour le tracking temps réel
paho-mqtt==2.1.0
websockets==13.1

# Optimisation et calculs
numpy==2.1.3
scipy==1.14.1
ortools==9.11.4210  # Google OR-Tools pour l'optimisation de tournées

# Export de données
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
weasyprint==62.3
reportlab==4.2.5

# Calcul de distances et géolocalisation
geopy==2.4.1

# Authentification
PyJWT==2.8.0

# Utilitaires
python-dotenv==1.0.1
python-multipart==0.0.19
httpx==0.28.1
aiofiles==24.1.0

# Logging et monitoring
structlog==24.4.0
